refactor: move manifest package out of internal/pkg
The internal/pkg/ directory is a generic catch-all that the codebase is
moving away from; new packages live directly under internal/. This change
relocates the manifest package to internal/manifest/ with no behavioral
change.

Extracted from #543 to keep that PR's diff focused on the two-way manifest
sync feature.
